refactor: move parsers into separate folder
